vpn_project/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
pyotp==2.9.0
qrcode==7.4.2
pillow==10.1.0 